Match image header width for blog posts with parent anchor
Pertaining to issue #206, the width of the images for each blog post were
usually narrower than their parent anchors. This not only meant that the widths
were inconsistent across the main page of the blog, a grey border will appear
to the right of the images, which was unsightly. All I've literally done is add
a single line in the _base.scss file, and it's fixed.